changed Float to Double

+++ src/main/kotlin/rmc/db/tables/RentalsTable.kt
@@ -9,10 +9,10 @@ object RentalsTable : IntIdTable() {
     val vehicleId = integer("vehicleId").references(VehiclesTable.id, onDelete = ReferenceOption.CASCADE)
     val userId = integer("userId").references(UsersTable.id, onDelete = ReferenceOption.CASCADE)
     val date = date("date")
-    val price = float("price").default(0.0F)
+    val price = double("price").default(0.00)
     val latitude = decimal("latitude", 8, 6)
     val longitude = decimal("longitude", 9, 6)
     val status = enumerationByName<RentalStatus>("status", 100).default(RentalStatus.PENDING)
-    val distanceTravelled = float("distanceTravelled").default(0.0F)
+    val distanceTravelled = double("distanceTravelled").default(0.00)
     val score = integer("score").default(100)
 }
